<WIP> 修改了人物转向的逻辑，根据鼠标滑动

+++ Assets/Scripts/Protocol/Message/World/FrameOptReqMsg.cs
@@ -7,13 +7,13 @@ using ZeroFrame.Buffer;namespace Game.Protocol.World
     public class FrameOptReqMsg:IZeroMessage<FrameOptReqMsg>{
         public int clientFrameIndex;
         public sbyte optTypeId;
-        public int msg;
+        public ulong msg;
 
         public void FromBytes(byte[] src, ref int offset)
         {
             clientFrameIndex = BufferReader.ReadInt32(src, ref offset);
             optTypeId = BufferReader.ReadSByte(src, ref offset);
-            msg = BufferReader.ReadInt32(src, ref offset);
+            msg = BufferReader.ReadUInt64(src, ref offset);
             offset += src.Length;
         }
 
@@ -23,7 +23,7 @@ using ZeroFrame.Buffer;namespace Game.Protocol.World
             byte[] result = new byte[1000];
             BufferWriter.WriteInt32(result, clientFrameIndex, ref offset);
             BufferWriter.WriteSByte(result, optTypeId, ref offset);
-            BufferWriter.WriteInt32(result, msg, ref offset);
+            BufferWriter.WriteUInt64(result, msg, ref offset);
             return result;
         }
